WIP. Simplifies schema list loading
